# Improve speed of read method. Improve speed of readinto. Fix readline method. Fix readable() method.