## Faster read()

in the same vein as #14 .


I removed all the checking from the read method. 

I tested it with:
```Python
#/usr/bin/env python3

import sys
import xopen

with xopen.xopen(sys.argv[1]) as file_h:
    block = True
    while block:
        block = file_h.read(4096)
```

Test results
before:
```
hyperfine -w 3 "python3 test.py ~/test/big.fastq.gz"
Benchmark #1: python3 test.py ~/test/big.fastq.gz
  Time (mean ± σ):      4.616 s ±  0.066 s    [User: 9.714 s, System: 1.690 s]
  Range (min … max):    4.519 s …  4.737 s    10 runs
 ```
After:
```
 hyperfine -w 3 "python3 test.py ~/test/big.fastq.gz"
Benchmark #1: python3 test.py ~/test/big.fastq.gz
  Time (mean ± σ):      4.617 s ±  0.069 s    [User: 8.666 s, System: 1.404 s]
  Range (min … max):    4.518 s …  4.705 s    10 runs
```
`User` equals total cpu time.
Before: 9.7 - 6.8 = 2.9 cpu seconds python overhead.
After: 8.7 - 6.8 = 1.9 cpu seconds python overhead. Pretty significant uplift.

## Fixing readline()
the original `readline()` implementation hanged. This reads one line, but then it waits for the process to finish. Process fills the kernel buffer for that certain process, and then the process stops until there is room in the buffer. `wait()` waits forever and `readline()` never returns data. This is fixed. Readline now works as expected.
This was probably missed in the tests if files smaller than 65 kb are used.

## Faster readinto
This is used in cutadapt. The variable assignment is skipped, this should save some (very small amount of) time. 

## Fix readable method.
This method existed in the PipedGzipReader class in python3 but not in python2. This is not considered best practice. It now always exists and raises a NotImplementedError in python2.