@rhpvorderman 2019-08-12
I see you do not have a `codecov.yml` yet in the repo. The requirements for a PR and the diff can be set this way for example: https://github.com/LUMC/fastqsplitter/blob/develop/codecov.yml. More documentation here: https://docs.codecov.io/docs/codecov-yaml. Only if you are not happy with the defaults of course! Just throwing it out there.