@marcelm 2019-08-12
I would assume that the gains are typically smaller for `read()` than your measurements show because I would expect a much bigger buffer size to be used. However, this change makes the code smaller, which is always good. Thanks for fixing `readline()`. If performance really matters, one could assign `self.readline = self._file.readline` in `__init__`, but this is somewhat obscure, so I’d prefer not to do it.

The difference for `readinto()` will not be measurable.

I don’t quite agree with your comment "this is not considered best practice." regarding `readable()` and wouldn’t consider the change you made a "fix", but I’m fine either way.

I’ve created a `codecov.yml` this morning but didn’t want to add it while your other PR was still open (just to avoid unnecessary merges).

Thanks again!