@rhpvorderman 2019-08-12
> I don’t quite agree with your comment "this is not considered best practice." regarding readable() and wouldn’t consider the change you made a "fix", but I’m fine either way.

This "best practice" stuff was a bit presumptuous on my part. Sorry for that! I heard in [this very nice talk](https://www.youtube.com/watch?v=DoWY7qQ9doA) that the getattr and setattr stuff in python was a bit unworkable and that now people aren't doing that so much anymore because it is not necessarily nice to work with changing methods. I agree with that and therefore I made the change, but I should not have pretended that I have the wisdom of "best practice"

> I would assume that the gains are typically smaller for read() than your measurements show because I would expect a much bigger buffer size to be used. 

Buffer sizes are a very interesting thing. Once they are too big they are not in the processor cache anymore, have to be transferred over the memory bus and that slows everything down. For writing I found once that the optimum for my particular processor was somewhere in the 100 fastq records (40000 bytes) ball park. That seems reasonable given that L1 and L2 cache are in the 32kb - 256kb ball park.  A 1000 reads (400kb) already had a significant slowdown (~25%) since L3 cache is already a lot slower. Have not tested reading yet.

Come to think of it. This slowdown might not have to do with the cache, but also with the kernel buffer which is 65 kb for these sort of things. Maybe the computer scientists know...

Thank you for merging and your constructive comments!