@marcelm 2019-08-13
It’s interesting you think the cache should have an effect. I was assuming up till now that these low-level effects don’t make much of a difference in Python because of the overhead of Python itself. But I’ll keep it in mind.